requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
email-validator
python-jose
passlib
python-multipart